+++ changelog.md
@@ -1,8 +1,18 @@
+\- **Version 3.02**  
+
+\* Deleted soft : Nmap , Udfhack , Mirai , Bom , Tool , Malware , IFC , Kawai Botnet , Darksploit , Rootnet
+
+\* Added soft : Shellphish , Vulnx , Email2PhoneNumber , Seeker , Inshackle , Gentiny , PyFlooder , TBomb , Spy-King , LittleBrother
+
+\* Deleted file "install.sh" because problems with compatibility 
+
+
+
 \- **Version 3.01**  
 
 \* Add install.sh for requirements
 
-\* Add Cam-Hacker to the list of soft
+\* Added soft : Cam-Hacker 
 
 \* Add files changelog.md